% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_top_n.R
\name{column_top_n}
\alias{column_top_n}
\title{Column Top n

Given a data and a column, this function counts the number of observations for each factor of the column levels and returns the top n with the most observations. If n is equal to the number of levels, then returns the number of observation for each level.}
\usage{
column_top_n(data, col, n)
}
\arguments{
\item{data}{is input data frame}

\item{n}{is the number of top levels to return}

\item{column}{is a column of data to be used for counting}
}
\value{
a table with columns column (given as argument in function) and c (the count of observation) for each n (top) row
}
\description{
Column Top n

Given a data and a column, this function counts the number of observations for each factor of the column levels and returns the top n with the most observations. If n is equal to the number of levels, then returns the number of observation for each level.
}
\examples{
column_top_n(datateachr::vancouver_trees, genus_name, 3)
column_top_n(datateachr::cancer_sample, diagnosis, 2)

}
